doc/html removed, this will be autogenerated by calling `make doc`. network support added, some of the network examples can be build now.
Tango BitArray fixed. tango.io.selector is working now. more examples can be compiled now.

gitignore and makefile for linux added

fixed some const correctness in tango/math/internal/BiguintCore.d

Adding tango.sql.Mysql

Mysql is compileable and works. A small example was also provided.
Extended a low level function in stringz

tango compileable on amd64 bit a lot of uints replaces by size_t

Fixed some octal bugs in tango/text/convert/Integer.d
Thread/Fiber class now derive from core.thread

doc/example/networking/* is now compileable on amd64.

tango.util.Convert.octal introduced. Some octal fixtures.

added some override statements in order to compile with -w flag.

core.bitop will now be used instead of tango/core/BitManip.d

fixed toString() method in tango.io.device.Conduit

tango.core.Thread constructor fixed.

hellophobos.d example added
some override statements added

add some override statements on tango.net.*.

rebase commit

tango.stdc replaced by core.stdc

Linux.mak renamed to posix.mak
ThreadExceptions are now public imported in tango.core.Thread;
Some case fallthrough statements fixed

tango.sys.Common; Fixed missing symbols.
tango.text.json.Jason converted to D2.

All unittests are should be compileable now.

Tango's default is MODEL=32

Tango's examples fixed

LocalSocket added to posix.mak

Base64 added to posix.mak

Fixed LocalSockets

+++ tango/io/stream/Buffered.d
@@ -425,7 +425,7 @@ class BufferedInput : InputFilter, InputBuffer
 
         ***********************************************************************/
 
-        final bool skip (int size)
+        final bool skip (size_t size)
         {
                 if (size < 0)
                    {
